books/requirements.txt
apturl==0.5.2
argcomplete==1.8.1
astroid==2.11.5
async-generator==1.10
attrs==21.4.0
autopep8==1.5.5
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certauth==1.1.6
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
click-config-file==0.6.0
click-plugins==1.1.1
clickhouse-driver==0.2.3
colorama==0.4.4
colorlog==6.6.0
command-not-found==0.3
configobj==5.0.6
cryptography==37.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.4
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
doublethink==0.3.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
flake8==3.8.4
future==0.18.2
gpg===1.16.0-unknown
gyp==0.1
h11==0.13.0
httplib2==0.20.2
humanize==4.1.0
idna==3.3
importlib-metadata==4.6.4
isort==5.10.1
jedi==0.17.2
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
nox==2022.1.7
oauthlib==3.2.0
olefile==0.46
outcome==1.1.0
packaging==21.3
PAM==0.4.2
paramiko==2.9.3
parso==0.7.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.21
pycups==2.0.1
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.11.2
PyGObject==3.42.0
PyICU==2.8.1
PyJWT==2.3.0
pylint==2.13.9
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rethinkdb==2.3.0.post6
rope==1.0.0
SecretStorage==3.3.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
trash-cli==0.17.1.14
trio==0.21.0
trio-websocket==0.9.2
twarc==2.10.4
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.2.0
unattended-upgrades==0.1
urlcanon==0.3.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
warcprox==2.4.17
warctools==4.10.0
wrapt==1.14.1
wsproto==1.1.0
xdg==5
xkit==0.0.0
yapf==0.32.0
zipp==1.0.0
